% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_car.R, R/geom_dog.R, R/geom_wick.R
\name{draw_key_wick}
\alias{draw_key_wick}
\title{creates the legend for the custom ggplot}
\usage{
draw_key_wick(data, size)

draw_key_wick(data, size)

draw_key_wick(data, size)
}
\arguments{
\item{data}{df that contains relevant info: the path and filename of the PNG image and size of image to be drawn to legend}

\item{size}{size of key realtive to overall plot used to scale the size of the image within the legend}
}
\description{
creates the legend for the custom ggplot

creates the legend for the custom ggplot

creates the legend for the custom ggplot
}
