% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_wick.R
\name{geom_wick}
\alias{geom_wick}
\title{geom_wick constructs a new layer for ggplot2 layer via the GeomWick for geom argument}
\usage{
geom_wick(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{aesthetic mapping what variables are mapped to x, y}

\item{data}{what data set to use}

\item{stat}{statistical transformation default is identity aka no transformation}

\item{position}{position for overlapping elements set to default}

\item{na.rm}{whether to remove NA values set to the default false}

\item{show.legend}{include or not the legend}

\item{inherit.aes}{set to true to inherit aesthetics from the global settings}
}
\description{
geom_wick constructs a new layer for ggplot2 layer via the GeomWick for geom argument
}
