% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_car.R
\docType{data}
\name{GeomCar}
\alias{GeomCar}
\title{GeomCar define the new geom}
\format{
An object of class \code{GeomCar} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomCar
}
\description{
GeomCar define the new geom
}
\keyword{datasets}
