% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dog.R
\docType{data}
\name{GeomDog}
\alias{GeomDog}
\title{GeomDog define the new geom}
\format{
An object of class \code{GeomDog} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomDog
}
\description{
GeomDog define the new geom
}
\keyword{datasets}
