% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_wick.R
\docType{data}
\name{GeomWick}
\alias{GeomWick}
\title{GeomWick define the new geom}
\format{
An object of class \code{GeomWick} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomWick
}
\description{
GeomWick define the new geom
}
\keyword{datasets}
